Step 5.2 - Added DomainPlugin for current platform, dispatching to legacy value formatter for AnnotationProxies
